Penguin v7 protocol
Penguin v7 spec WIP

Implement frame types

Go back to 16-bit stream ID

Frame changes for 16bit stream id

Finish frame

Cosmetic changes

Enable builds on AtomicU64-absent platforms

Fix docs

Protocol redesign

New protocol spec

Finish porting old mux functions over

Update old code for the new proto

Fix testing env

+++ src/mux/config.rs
@@ -10,8 +10,8 @@ pub const DATAGRAM_BUFFER_SIZE: usize = 1 << 9;
 /// need to have a crazy high buffer size.
 pub const STREAM_BUFFER_SIZE: usize = 1 << 5;
 
-/// Needs to be the same as `STREAM_FRAME_BUFFER_SIZE` but as `u64`
-pub const RWND: u64 = STREAM_FRAME_BUFFER_SIZE as u64;
+/// Needs to be the same as `STREAM_FRAME_BUFFER_SIZE` but as `u32`
+pub const RWND: u32 = STREAM_FRAME_BUFFER_SIZE as u32;
 
 /// Number of `StreamFrame`s to buffer in `MuxStream`'s channels before blocking
 #[cfg(not(test))]
@@ -20,7 +20,7 @@ pub const STREAM_FRAME_BUFFER_SIZE: usize = 1 << 9;
 /// If too low, `Ack`s will consume too much bandwidth;
 /// If too high, writers may block.
 #[cfg(not(test))]
-pub const DEFAULT_RWND_THRESHOLD: u64 = 1 << 8;
+pub const DEFAULT_RWND_THRESHOLD: u32 = 1 << 8;
 
 /// Number of `StreamFrame`s to buffer in `MuxStream`'s channels before blocking
 #[cfg(test)]
@@ -29,4 +29,4 @@ pub const STREAM_FRAME_BUFFER_SIZE: usize = 1 << 2;
 /// test the `Ack` mechanism, so we set this to be the same as the buffer size.
 /// The downside is that tests will be slower.
 #[cfg(test)]
-pub const DEFAULT_RWND_THRESHOLD: u64 = RWND;
+pub const DEFAULT_RWND_THRESHOLD: u32 = RWND;
